requirements.txt
Django>=3.2.4,<3.3                      # Between 3.2 and 3.3 to get 3.2 later releases. Core django
djangorestframework>=3.12.4,<3.13       # Between 3.12 and 3.13 to get 3.12 later releass. Rest api
psycopg2>=2.8.6,<2.9                    # Between 2.8.6 and 3.13 to get 2.9 later releass. Postgres helper
drf-spectacular>=0.15.1,<0.16           # Between 0.15.1 and 3.13 to get 0.16 later releass. Open api helper
djangorestframework-simplejwt==5.2.0    # JWT helper
python-decouple==3.6                    # Config helper
django-redis>=5.2.0,<5.3                # Between 5.2.0 and 5.3 to get 0.16 later releass. Cache, async task helper
Pillow>=8.2.0,<8.3.0                    # Between 8.2.0 and 8.3.0 to get later releass. Image file helper